Remove unnecessary tool iscsiuio
iscsiuio is used in conjunction with specific linux driver to
improve performance, such as QLogic NetXtreme II or QLogic CNIC driver.
It is not a necessary tool of open-iscsi.

What's more, iscsiuio used a package uio which is out of maintain now,
which would introduce CVEs unhandled. So I want to remove this tool
from open-iscsi.

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>